refactor: paginatedResult/PaginationOptions を shared に移動
employee サブドメイン固有だった型を共通化し、他のサブドメインからも利用可能にする。
